# docs/swagger.yaml
basePath: /api/v1
definitions:
  dto.ApiKeyResponse:
    properties:
      api_key:
        type: string
    type: object
  dto.ChangePasswordRequest:
    properties:
      new_password:
        example: 1Pass123!@#
        type: string
      old_password:
        example: Pass123!@#
        type: string
    required:
    - new_password
    - old_password
    type: object
  dto.DeleteRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  dto.ForgotPasswordRequest:
    properties:
      email:
        example: john@example.com
        type: string
      otp:
        type: string
      password:
        example: Password1!@#
        minLength: 8
        type: string
    required:
    - email
    - otp
    - password
    type: object
  dto.LoginRequest:
    properties:
      email:
        example: john@example.com
        type: string
      password:
        example: Pass123!@#
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
  dto.LoginResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    required:
    - access_token
    - refresh_token
    type: object
  dto.LogoutRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  dto.PageCaptureRequest:
    properties:
      delaySeconds:
        type: integer
      fullPage:
        type: boolean
      height:
        type: integer
      isMobile:
        type: boolean
      url:
        type: string
      width:
        type: integer
    required:
    - url
    type: object
  dto.RefreshTokenRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  dto.RefreshTokenResponse:
    properties:
      access_token:
        type: string
    required:
    - access_token
    type: object
  dto.RegisterRequest:
    properties:
      confirm:
        example: Pass123!@#
        minLength: 8
        type: string
      email:
        example: john@example.com
        type: string
      name:
        example: John Doe
        type: string
      password:
        example: Pass123!@#
        minLength: 8
        type: string
    required:
    - confirm
    - email
    - name
    - password
    type: object
  dto.SendOtpRequest:
    properties:
      email:
        example: john@example.com
        type: string
    required:
    - email
    type: object
  dto.SetPasswordRequest:
    properties:
      confirm:
        example: Password1!@#
        minLength: 8
        type: string
      password:
        example: Password1!@#
        minLength: 8
        type: string
    required:
    - confirm
    - password
    type: object
  dto.UserResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      email_verified:
        type: string
      name:
        type: string
      profile_picture:
        type: string
      updated_at:
        type: string
      uuid:
        type: string
    type: object
  dto.VerifyEmailRequest:
    properties:
      email:
        example: john@example.com
        type: string
      otp:
        type: string
    required:
    - email
    - otp
    type: object
  entity.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      email_verified:
        type: string
      name:
        type: string
      password:
        type: string
      profile_picture:
        type: string
      public_id:
        type: string
      updated_at:
        type: string
      uuid:
        type: string
    type: object
  response.ErrorResponse:
    properties:
      error:
        type: string
      message:
        type: string
      status:
        type: integer
    type: object
  response.Response:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      status:
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is a documentation for PageCapApi
  title: Page Capture API (PageCapApi)
  version: "1.0"
paths:
  /auth/api-key:
    get:
      consumes:
      - application/json
      description: Generate an API key for user
      produces:
      - application/json
      responses:
        "200":
          description: Successfully generate API key
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.ApiKeyResponse'
              type: object
        "400":
          description: invalid request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Generate API Key
      tags:
      - Auth
  /auth/forgot-password:
    post:
      consumes:
      - application/json
      description: Reset password using OTP
      parameters:
      - description: Forgot Password Request
        in: body
        name: forgotPassword
        required: true
        schema:
          $ref: '#/definitions/dto.ForgotPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Password reset successfully
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: invalid request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Forgot password
      tags:
      - Auth
  /auth/google:
    get:
      consumes:
      - application/json
      description: Login using Google OAuth2
      produces:
      - application/json
      responses:
        "200":
          description: Google login successful
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: invalid request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Google login
      tags:
      - Auth
  /auth/google/callback:
    get:
      consumes:
      - application/json
      description: Callback URL for Google OAuth2
      produces:
      - application/json
      responses:
        "200":
          description: Google login successful
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.LoginResponse'
              type: object
        "400":
          description: invalid request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Google callback
      tags:
      - Auth
  /auth/login:
    post:
      consumes:
      - application/json
      description: Login user with email and password
      parameters:
      - description: Login Request
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "201":
          description: user logged in successfully
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.LoginResponse'
              type: object
        "400":
          description: invalid request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Login user
      tags:
      - Auth
  /auth/logout:
    delete:
      consumes:
      - application/json
      description: Logout user and invalidate the refresh token
      parameters:
      - description: Logout Request
        in: body
        name: logout
        required: true
        schema:
          $ref: '#/definitions/dto.LogoutRequest'
      produces:
      - application/json
      responses:
        "200":
          description: user logged out successfully
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: invalid request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Logout user
      tags:
      - Auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Refresh access token using refresh token
      parameters:
      - description: Refresh Token Request
        in: body
        name: refresh
        required: true
        schema:
          $ref: '#/definitions/dto.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: access token refreshed successfully
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.RefreshTokenResponse'
              type: object
        "400":
          description: invalid request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Refresh access token
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user with name, email, and password
      parameters:
      - description: Register Request
        in: body
        name: register
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: user registered successfully
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/entity.User'
              type: object
        "400":
          description: invalid request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Register a new user
      tags:
      - Auth
  /auth/send-otp:
    post:
      consumes:
      - application/json
      description: Send OTP to user's email
      parameters:
      - description: Send OTP Request
        in: body
        name: sendOtp
        required: true
        schema:
          $ref: '#/definitions/dto.SendOtpRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OTP sent successfully
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: invalid request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Send OTP
      tags:
      - Auth
  /auth/set-password:
    post:
      consumes:
      - application/json
      description: Set password for the user after Google login
      parameters:
      - description: Set Password Request
        in: body
        name: setPassword
        required: true
        schema:
          $ref: '#/definitions/dto.SetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Password set successfully
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: invalid request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Set password
      tags:
      - Auth
  /auth/verify-email:
    post:
      consumes:
      - application/json
      description: Verify email using OTP
      parameters:
      - description: Verify Email Request
        in: body
        name: verifyEmail
        required: true
        schema:
          $ref: '#/definitions/dto.VerifyEmailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Email verified successfully
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: invalid request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Verify email
      tags:
      - Auth
  /page-capture/{key}:
    post:
      consumes:
      - application/json
      description: Get Page Capture
      parameters:
      - description: Key for Page Capture
        in: path
        name: key
        required: true
        type: string
      - description: Page Capture Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.PageCaptureRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully get Page Capture image
          schema:
            type: file
        "400":
          description: invalid request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get Page Capture
      tags:
      - Page Capture
  /user:
    delete:
      consumes:
      - application/json
      description: Delete user account
      parameters:
      - description: Delete User Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.DeleteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully delete user account
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: invalid request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete User
      tags:
      - User
  /user/change-password:
    patch:
      consumes:
      - application/json
      description: Change user password
      parameters:
      - description: Change Password Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully change user password
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.UserResponse'
              type: object
        "400":
          description: invalid request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Change Password
      tags:
      - User
  /user/profile:
    get:
      consumes:
      - application/json
      description: Get user profile
      produces:
      - application/json
      responses:
        "200":
          description: Successfully get user profile
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.UserResponse'
              type: object
        "400":
          description: invalid request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get Profile
      tags:
      - User
    patch:
      consumes:
      - multipart/form-data
      description: Update user profile
      parameters:
      - description: Profile file
        in: formData
        name: profile_picture
        required: true
        type: file
      - description: User name
        in: formData
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.UserResponse'
              type: object
        "400":
          description: invalid request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update User Profile
      tags:
      - User
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"